removed graph calls back to their repsective components
