+++ components/Repo/Table.js
@@ -2,6 +2,7 @@ import React, { Component } from 'react'
 import { compose } from 'redux'
 import { css } from 'glamor'
 import { graphql } from 'react-apollo'
+import gql from 'graphql-tag'
 import { descending, ascending } from 'd3-array'
 
 import withT from '../../lib/withT'
@@ -25,8 +26,6 @@ import { Table, Tr, Th, ThOrder, Td, TdNum } from '../Table'
 import Loader from '../Loader'
 
 import { GITHUB_ORG, REPO_PREFIX, FRONTEND_BASE_URL } from '../../lib/settings'
-import { filterAndOrderRepos } from '../../lib/graphql/queries'
-import { editRepoMeta } from '../../lib/graphql/mutations'
 
 import {
   matchType
@@ -38,6 +37,88 @@ import Briefing from './Briefing'
 import EditMetaDate from './EditMetaDate'
 import { phases } from './workflow'
 import RepoAdd from './Add'
+
+export const editRepoMeta = gql`
+mutation editRepoMeta(
+  $repoId: ID!
+  $creationDeadline: DateTime
+  $productionDeadline: DateTime
+  $publishDate: DateTime
+  $briefingUrl: String
+) {
+  editRepoMeta(
+    repoId: $repoId
+    creationDeadline: $creationDeadline
+    productionDeadline: $productionDeadline
+    publishDate: $publishDate
+    briefingUrl: $briefingUrl
+  ) {
+    id
+    meta {
+      creationDeadline
+      productionDeadline
+      publishDate
+      briefingUrl
+    }
+  }
+}
+`
+
+export const filterAndOrderRepos = gql`
+query repoListSearch($after: String, $search: String, $orderBy: RepoOrderBy) {
+  repos(
+    first: 100,
+    after: $after,
+    search: $search,
+    orderBy: $orderBy
+  ) {
+    totalCount
+    pageInfo {
+      endCursor
+      hasNextPage
+    }
+    nodes {
+      id
+      meta {
+        creationDeadline
+        productionDeadline
+        publishDate
+        briefingUrl
+      }
+      latestCommit {
+        id
+        date
+        message
+        document {
+          id
+          meta {
+            template
+            title
+            credits
+          }
+        }
+      }
+      milestones {
+        name
+        immutable
+      }
+      latestPublications {
+        name
+        prepublication
+        live
+        scheduledAt
+        document {
+          id
+          meta {
+            path
+            slug
+          }
+        }
+      }
+    }
+  }
+}
+`
 
 const dateTimeFormat = '%d.%m %H:%M'
 const formatDateTime = swissTime.format(dateTimeFormat)
