--- lib/graphql/index.js
@@ -1,6 +1,0 @@
-import * as fragments from './fragments'
-import * as mutations from './mutations'
-import * as subscriptions from './subscriptions'
-import * as queries from './queries'
-
-export { fragments, mutations, subscriptions, queries }
